Reduce redundant cast warnings. These happen as a side result of code cleanup..
Reduce warnings that look like this:

   [repeat] /home/bwalker/src/netbeans/enterprise/tomcat5/src/org/netbeans/modules/tomcat5/util/LogManager.java:271: warning: [cast] redundant cast to LogViewer
   [repeat]         contextLog = (LogViewer)contextLogViewers.get(moduleID);
   [repeat]                      ^Cleanup redundant casts. These are a side effect of work done doing code cleanup.

+++ enterprise/jakarta.web.beans/src/org/netbeans/modules/jakarta/web/beans/completion/CCPaintComponent.java
@@ -215,7 +215,7 @@ public class CCPaintComponent extends JPanel {
     }
     
     protected int getWidth(String s) {
-        Integer i = (Integer)widths.get(s);
+        Integer i = widths.get(s);
         if (i != null) {
             return i;
         } else {
